requirements.txt
chromadb
langchain
langchain-text-splitters
langchain-community
langchain-openai
langchain-core
langchain-anthropic
langchain-mistralai
langgraph
pydantic
pdfplumber
python-dotenv
requests
streamlit
grandalf
nltk
matplotlib
sentence_transformers
retry